Add PR publishing workflow generation

+++ src/main/groovy/net/neoforged/gradleutils/tasks/CIConfigExtractionTask.java
@@ -8,10 +8,12 @@ package net.neoforged.gradleutils.tasks;
 import org.gradle.api.DefaultTask;
 import org.gradle.api.file.DirectoryProperty;
 import org.gradle.api.plugins.JavaPluginExtension;
+import org.gradle.api.tasks.Input;
 import org.gradle.api.tasks.InputDirectory;
 import org.gradle.api.tasks.PathSensitive;
 import org.gradle.api.tasks.PathSensitivity;
 import org.gradle.api.tasks.TaskAction;
+import org.gradle.api.tasks.options.Option;
 import org.gradle.jvm.toolchain.JavaLanguageVersion;
 
 import java.io.File;
@@ -19,6 +21,9 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.util.ArrayDeque;
+import java.util.Deque;
+import java.util.List;
 import java.util.zip.ZipEntry;
 
 public abstract class CIConfigExtractionTask extends DefaultTask {
@@ -42,7 +47,7 @@ public abstract class CIConfigExtractionTask extends DefaultTask {
         String fileZip = exportResource();
 
         //Check if the directory exists, if so then delete it.
-        if (targetDir.exists()) {
+        if (targetDir.exists() && !disableDeletion) {
             boolean couldDelete = true;
             for (File file : targetDir.listFiles()) {
                 couldDelete &= file.delete();
@@ -124,5 +129,41 @@ public abstract class CIConfigExtractionTask extends DefaultTask {
         }
 
         return jarFolder + templateZipName;
+    }
+
+    private boolean disableDeletion;
+
+    @Input
+    public boolean getDisableDeletion() {
+        return this.disableDeletion;
+    }
+
+    @Option(option = "disable-deletion", description = "Disable deletion of existing workflows")
+    public void setDisableDeletion(boolean disableDeletion) {
+        this.disableDeletion = disableDeletion;
+    }
+
+    /**
+     * Finds the most common 0-based substring of all {@code groups}.
+     */
+    protected static String findCommonSubstring(List<String> groups) {
+        if (groups.isEmpty()) {
+            return "";
+        }
+
+        final Deque<Character> firstGroup = new ArrayDeque<>();
+        groups.get(0).chars().forEach(ch -> firstGroup.add(Character.valueOf((char) ch)));
+
+        String common = "";
+        while (!firstGroup.isEmpty()) {
+            final String current = common + firstGroup.pop();
+            if (groups.stream().allMatch(group -> group.startsWith(current))) {
+                common = current;
+            } else {
+                break;
+            }
+        }
+
+        return common;
     }
 }
